perf: 优化commitlog和consumequeue文件写入体积

+++ mqbroker/src/main/java/com/k/mq/broker/model/CommitLogModel.java
@@ -1,24 +1,30 @@
 package com.k.mq.broker.model;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * CommitLog模型
  * 表示CommitLog文件的配置和状态信息
- * 
+ *
  * @author yihang07
  */
 public class CommitLogModel {
-    /** CommitLog文件名 */
+    /**
+     * CommitLog文件名
+     */
     private String fileName;
-    /** 偏移量上限（字节数） */
+    /**
+     * 偏移量上限（字节数）
+     */
     private Long offsetLimit;
-    /** 当前偏移量（原子操作，支持并发） */
-    private AtomicLong offset;
+    /**
+     * 当前偏移量（原子操作，支持并发）
+     */
+    private AtomicInteger offset;
 
     /**
      * 获取文件名
-     * 
+     *
      * @return 文件名
      */
     public String getFileName() {
@@ -27,7 +33,7 @@ public class CommitLogModel {
 
     /**
      * 设置文件名
-     * 
+     *
      * @param fileName 文件名
      */
     public void setFileName(String fileName) {
@@ -36,7 +42,7 @@ public class CommitLogModel {
 
     /**
      * 获取偏移量上限
-     * 
+     *
      * @return 偏移量上限
      */
     public Long getOffsetLimit() {
@@ -45,7 +51,7 @@ public class CommitLogModel {
 
     /**
      * 设置偏移量上限
-     * 
+     *
      * @param offsetLimit 偏移量上限
      */
     public void setOffsetLimit(Long offsetLimit) {
@@ -54,25 +60,25 @@ public class CommitLogModel {
 
     /**
      * 获取当前偏移量（原子类型）
-     * 
+     *
      * @return 当前偏移量的AtomicLong对象
      */
-    public AtomicLong getOffset() {
+    public AtomicInteger getOffset() {
         return offset;
     }
 
     /**
      * 设置当前偏移量
-     * 
+     *
      * @param offset 当前偏移量的AtomicLong对象
      */
-    public void setOffset(AtomicLong offset) {
+    public void setOffset(AtomicInteger offset) {
         this.offset = offset;
     }
 
     /**
      * 计算剩余可用空间
-     * 
+     *
      * @return 偏移量上限与当前偏移量的差值（剩余空间）
      */
     public Long diff() {
